user-service/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
click==8.3.0
cloudinary==1.44.1
Deprecated==1.2.18
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.117.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
limits==5.5.0
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
